test2

+++ lab02/lab02.c
@@ -1,23 +1,21 @@
 #include <stdio.h>
 #include <math.h>
-#include <stdlib.h>
 
 int main()
 {
     double x, y, s;
     x = 0.0; 
-    scanf("%lf", &s);
-    int n = 2 / s + 1;
+    scanf("%lf", &s); 
+    double n = 2 / s + 1;
     while (n) 
    {
         if (x >= -1 && x <= 1)
           y = exp(-2*sin(x));
         if (x > 1 && x <= 2)
           y = x*x - cos(x)/sin(x);
-         printf("%f %f\n", x, y);
+         printf("%lf %lf\n", x, y);
         x += s;
         n = n-1;
    }
 
 }
-
